feat: sessions DB table + dual graph views (Step Graph + Topology)
Sessions DB:
- New sessions table (auto-migrated on backend startup)
- Composite index (namespace, updated_at DESC) for fast listing
- Atomic upsert on /chat/stream with COALESCE for safe overwrites
- Prepared fields: owner_email, owner_sub, model_override, budget limits

Graph Views:
- StepGraphView: restored old per-step graph
- TopologyGraphView: topology DAG with edge descriptions
- GraphLoopView: wrapper with Step Graph / Topology subtab toggle
- Both accept allLoops for multi-message rendering
